parsers/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
idna==3.4
nats-py==2.6.0
redis==5.0.1
requests==2.31.0
urllib3==2.0.7
